Identity key change handling
When an incoming idenity key change is accepted, the old key is wiped,
the associated imcoming message is decrypted and has its contents
updated, and the error is removed.

To facilitate this process, the message uid (which can be used to fetch
the saved message from the database) is passed along with the decrypted
content to update records in the database.